Begin refactor of mobile patient view
